Separate Preferences class into two separated so that each has own scope
